+++ timeline/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-from .models import ProjectEvent, ProjectEventObjectRef, ProjectEventStatus
-
-
-admin.site.register(ProjectEvent)
-admin.site.register(ProjectEventObjectRef)
-admin.site.register(ProjectEventStatus)